requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# AI/ML Libraries
transformers==4.35.2
torch==2.2.0
openai==1.3.7
huggingface-hub==0.19.4

# Text processing
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3
langchain==0.0.340
faiss-cpu==1.7.4

# HTTP and API
requests==2.31.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Background tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Additional AI/ML tools
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Text analysis
textblob==0.17.1
vaderSentiment==3.3.2

# API documentation
python-multipart==0.0.6

# CORS and middleware
python-jose[cryptography]==3.3.0 